% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/or_95ci.R
\name{OR_95CI}
\alias{OR_95CI}
\title{Odds Ratios and 95\% Confidence Intervals from Logistic Regression Results}
\usage{
OR_95CI(coef, se, siglevel, roundto)
}
\arguments{
\item{coef}{numeric vector}

\item{se}{numeric vector equal in length to coef}

\item{siglevel}{the value of alpha, or the type 1 error rate, such that the output produces a 100(1 - siglevel)\% CI.}

\item{roundto}{the amount of digits after the decimal point to round each value to.}
}
\value{
a string containing the computed point estimates and SEs.
}
\description{
A function that returns the point estimate and 95\% CI of the Odds Ratios corresponding to an inputted set of coefficients and standard errors from logistic regression.
}
\examples{
OR_95CI(0, 1, 0.95, 3)
}
\author{
Daniel Del Rosso
}
