% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toydata_script.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{Example data}
\format{
A data frame with 100 rows and 3 variables:
\describe{
\item{Y}{Child satisfaction: a binary outcome}
\item{x1}{Rating Z-score: a continuous predictor}
\item{x2}{Educational designation: a binary predictor}
}
}
\usage{
toydata
}
\description{
Example data
}
\keyword{datasets}
